+++ tencentcloud/cpdp/v20190820/models.go
@@ -18887,26 +18887,30 @@ type QueryExternalAccountBookResult struct {
 
 // Predefined struct for user
 type QueryFinancialDataUrlRequestParams struct {
-	// 数据查询范围:结束时间
+	// 数据查询范围:结束时间 yyyy-MM-dd HH:mm:ss
 	EndTime *string `json:"EndTime,omitempty" name:"EndTime"`
 
-	// 数据查询范围:开始时间
+	// 数据查询范围:开始时间 yyyy-MM-dd HH:mm:ss
 	StartTime *string `json:"StartTime,omitempty" name:"StartTime"`
 
-	// 数据类型：ADDED_INVOICE_REPORT  增值税开票数据，NATURAL_FINANCE_REPORT 自然人金融数据
+	// 数据类型：ADDED_INVOICE_REPORT  增值税开票数据，NATURAL_FINANCE_REPORT 自然人金融数据,
+	// NATURAL_FINANCE_REPORT_BY_TIME 按次报税数据导出,
+	// NATURAL_FINANCE_REPORT_STAT_BY_DAY 按日累计报税数据导出
 	DataType *string `json:"DataType,omitempty" name:"DataType"`
 }
 
 type QueryFinancialDataUrlRequest struct {
 	*tchttp.BaseRequest
 	
-	// 数据查询范围:结束时间
+	// 数据查询范围:结束时间 yyyy-MM-dd HH:mm:ss
 	EndTime *string `json:"EndTime,omitempty" name:"EndTime"`
 
-	// 数据查询范围:开始时间
+	// 数据查询范围:开始时间 yyyy-MM-dd HH:mm:ss
 	StartTime *string `json:"StartTime,omitempty" name:"StartTime"`
 
-	// 数据类型：ADDED_INVOICE_REPORT  增值税开票数据，NATURAL_FINANCE_REPORT 自然人金融数据
+	// 数据类型：ADDED_INVOICE_REPORT  增值税开票数据，NATURAL_FINANCE_REPORT 自然人金融数据,
+	// NATURAL_FINANCE_REPORT_BY_TIME 按次报税数据导出,
+	// NATURAL_FINANCE_REPORT_STAT_BY_DAY 按日累计报税数据导出
 	DataType *string `json:"DataType,omitempty" name:"DataType"`
 }