+++ tencentcloud/dts/v20211206/client.go
@@ -2353,6 +2353,122 @@ func (c *Client) ResumeSyncJobWithContext(ctx context.Context, request *ResumeSy
     return
 }
 
+func NewSkipCheckItemRequest() (request *SkipCheckItemRequest) {
+    request = &SkipCheckItemRequest{
+        BaseRequest: &tchttp.BaseRequest{},
+    }
+    
+    request.Init().WithApiInfo("dts", APIVersion, "SkipCheckItem")
+    
+    
+    return
+}
+
+func NewSkipCheckItemResponse() (response *SkipCheckItemResponse) {
+    response = &SkipCheckItemResponse{
+        BaseResponse: &tchttp.BaseResponse{},
+    }
+    return
+}
+
+// SkipCheckItem
+// 本接口用于校验检查项不通过后，可进行跳过此校验项操作，后端将不再校验该项。任何校验步骤都是不应该跳过的，通过校验是能正确执行的前置条件。支持跳过的产品及链路的校验项可 [参考文档](https://cloud.tencent.com/document/product/571/61639)。
+//
+// 可能返回的错误码:
+//  INVALIDPARAMETERVALUE_INVALIDPARAMETERVALUEERROR = "InvalidParameterValue.InvalidParameterValueError"
+//  OPERATIONDENIED_OPERATIONDENIEDERROR = "OperationDenied.OperationDeniedError"
+func (c *Client) SkipCheckItem(request *SkipCheckItemRequest) (response *SkipCheckItemResponse, err error) {
+    return c.SkipCheckItemWithContext(context.Background(), request)
+}
+
+// SkipCheckItem
+// 本接口用于校验检查项不通过后，可进行跳过此校验项操作，后端将不再校验该项。任何校验步骤都是不应该跳过的，通过校验是能正确执行的前置条件。支持跳过的产品及链路的校验项可 [参考文档](https://cloud.tencent.com/document/product/571/61639)。
+//
+// 可能返回的错误码:
+//  INVALIDPARAMETERVALUE_INVALIDPARAMETERVALUEERROR = "InvalidParameterValue.InvalidParameterValueError"
+//  OPERATIONDENIED_OPERATIONDENIEDERROR = "OperationDenied.OperationDeniedError"
+func (c *Client) SkipCheckItemWithContext(ctx context.Context, request *SkipCheckItemRequest) (response *SkipCheckItemResponse, err error) {
+    if request == nil {
+        request = NewSkipCheckItemRequest()
+    }
+    
+    if c.GetCredential() == nil {
+        return nil, errors.New("SkipCheckItem require credential")
+    }
+
+    request.SetContext(ctx)
+    
+    response = NewSkipCheckItemResponse()
+    err = c.Send(request, response)
+    return
+}
+
+func NewSkipSyncCheckItemRequest() (request *SkipSyncCheckItemRequest) {
+    request = &SkipSyncCheckItemRequest{
+        BaseRequest: &tchttp.BaseRequest{},
+    }
+    
+    request.Init().WithApiInfo("dts", APIVersion, "SkipSyncCheckItem")
+    
+    
+    return
+}
+
+func NewSkipSyncCheckItemResponse() (response *SkipSyncCheckItemResponse) {
+    response = &SkipSyncCheckItemResponse{
+        BaseResponse: &tchttp.BaseResponse{},
+    }
+    return
+}
+
+// SkipSyncCheckItem
+// 本接口用于校验检查项不通过后，可进行跳过此校验项操作，后端将不再校验该项。任何校验步骤都是不应该跳过的，通过校验是能正确执行的前置条件。支持跳过的产品及链路的校验项可 [参考文档](https://cloud.tencent.com/document/product/571/61639)。
+//
+// 可能返回的错误码:
+//  AUTHFAILURE_AUTHFAILUREERROR = "AuthFailure.AuthFailureError"
+//  FAILEDOPERATION = "FailedOperation"
+//  INTERNALERROR = "InternalError"
+//  INTERNALERROR_DATABASEERROR = "InternalError.DatabaseError"
+//  INTERNALERROR_UNKNOWNERROR = "InternalError.UnknownError"
+//  INVALIDPARAMETER = "InvalidParameter"
+//  INVALIDPARAMETERVALUE = "InvalidParameterValue"
+//  OPERATIONDENIED = "OperationDenied"
+//  OPERATIONDENIED_OPERATIONDENIEDERROR = "OperationDenied.OperationDeniedError"
+//  RESOURCENOTFOUND = "ResourceNotFound"
+func (c *Client) SkipSyncCheckItem(request *SkipSyncCheckItemRequest) (response *SkipSyncCheckItemResponse, err error) {
+    return c.SkipSyncCheckItemWithContext(context.Background(), request)
+}
+
+// SkipSyncCheckItem
+// 本接口用于校验检查项不通过后，可进行跳过此校验项操作，后端将不再校验该项。任何校验步骤都是不应该跳过的，通过校验是能正确执行的前置条件。支持跳过的产品及链路的校验项可 [参考文档](https://cloud.tencent.com/document/product/571/61639)。
+//
+// 可能返回的错误码:
+//  AUTHFAILURE_AUTHFAILUREERROR = "AuthFailure.AuthFailureError"
+//  FAILEDOPERATION = "FailedOperation"
+//  INTERNALERROR = "InternalError"
+//  INTERNALERROR_DATABASEERROR = "InternalError.DatabaseError"
+//  INTERNALERROR_UNKNOWNERROR = "InternalError.UnknownError"
+//  INVALIDPARAMETER = "InvalidParameter"
+//  INVALIDPARAMETERVALUE = "InvalidParameterValue"
+//  OPERATIONDENIED = "OperationDenied"
+//  OPERATIONDENIED_OPERATIONDENIEDERROR = "OperationDenied.OperationDeniedError"
+//  RESOURCENOTFOUND = "ResourceNotFound"
+func (c *Client) SkipSyncCheckItemWithContext(ctx context.Context, request *SkipSyncCheckItemRequest) (response *SkipSyncCheckItemResponse, err error) {
+    if request == nil {
+        request = NewSkipSyncCheckItemRequest()
+    }
+    
+    if c.GetCredential() == nil {
+        return nil, errors.New("SkipSyncCheckItem require credential")
+    }
+
+    request.SetContext(ctx)
+    
+    response = NewSkipSyncCheckItemResponse()
+    err = c.Send(request, response)
+    return
+}
+
 func NewStartCompareRequest() (request *StartCompareRequest) {
     request = &StartCompareRequest{
         BaseRequest: &tchttp.BaseRequest{},